feat: add calculator button IDs, a decimal-point button, and fix precision (#39)
Add stable element IDs to the operator buttons and two new decimal-point
buttons so the calculator screen can be targeted by the Playwright suite,
matching the id-based selector convention already used elsewhere (e.g.
#done-list, #open-list).

The number inputs were type="number", which silently discards any value
ending in a bare "." (e.g. "3." is not a spec-valid number), so the
decimal button — and manual decimal entry on iOS, whose numeric keypad
for type="number" has no "." key — had no way to work. Switch them to
type="text" inputmode="decimal", which keeps the numeric keypad on
mobile while allowing intermediate values like "3." to survive.

Also switch the SOAP service's Add/Subtract/Multiply/Divide from Float
(32-bit) to Double (64-bit) so results match what a real calculator
would show for more precise inputs.

+++ api/soap_calculator.py
@@ -1,23 +1,23 @@
-from spyne import Application, Fault, Float, ServiceBase, rpc
+from spyne import Application, Double, Fault, ServiceBase, rpc
 from spyne.protocol.soap import Soap11
 from spyne.server.wsgi import WsgiApplication
 from wsgiref.simple_server import make_server
 
 
 class CalculatorService(ServiceBase):
-    @rpc(Float, Float, _returns=Float)
+    @rpc(Double, Double, _returns=Double)
     def Add(ctx, a, b):
         return a + b
 
-    @rpc(Float, Float, _returns=Float)
+    @rpc(Double, Double, _returns=Double)
     def Subtract(ctx, a, b):
         return a - b
 
-    @rpc(Float, Float, _returns=Float)
+    @rpc(Double, Double, _returns=Double)
     def Multiply(ctx, a, b):
         return a * b
 
-    @rpc(Float, Float, _returns=Float)
+    @rpc(Double, Double, _returns=Double)
     def Divide(ctx, a, b):
         if b == 0:
             raise Fault(faultcode="Client", faultstring="Division by zero is not allowed")
